// src/dealer/dealer.dto.ts
export class DealerDto {
    dealer_code: string;    
    dealer_name: string;
    nid : number;
    contact : string;
    address: string;
    status: string;   
    company_type: number;
    region: number;
    area: number;
    territory: number;
    cluster: number;
  }

  export class UpdateDealerDto {
    dealer_code: string;    
    dealer_name: string;
    nid : number;
    contact : string;
    address: string;
    status: string;   
    company_type: number;
    region: number;
    area: number;
    territory: number;
    cluster: number;
  }